account for dj_helper_fn move to utils
