+++ src/spyglass/position_v1/position_dlc_cohort.py
@@ -1,5 +1,5 @@
 import datajoint as dj
-from ..common.dj_helper_fn import fetch_nwb
+from ..utils.dj_helper_fn import fetch_nwb
 from ..common.common_nwbfile import AnalysisNwbfile
 from .position_dlc_pose_estimation import DLCPoseEstimation
 from .position_dlc_position import DLCSmoothInterp
