+++ src/spyglass/position_v1/position_dlc_pose_estimation.py
@@ -7,7 +7,7 @@ import numpy as np
 import cv2
 import datajoint as dj
 from IPython.display import display
-from ..common.dj_helper_fn import fetch_nwb
+from ..utils.dj_helper_fn import fetch_nwb
 from ..common.common_behav import VideoFile, RawPosition
 from ..common.common_nwbfile import AnalysisNwbfile
 from .position_dlc_project import BodyPart
