+++ src/spyglass/position_v1/position_dlc_position.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 import datajoint as dj
-from ..common.dj_helper_fn import fetch_nwb
+from ..utils.dj_helper_fn import fetch_nwb
 from ..common.common_nwbfile import AnalysisNwbfile
 from ..common.common_behav import RawPosition
 from .position_dlc_pose_estimation import DLCPoseEstimation
