+++ src/spyglass/position_v1/position_dlc_selection.py
@@ -6,7 +6,7 @@ import datajoint as dj
 from tqdm import tqdm as tqdm
 import pynwb
 import cv2
-from ..common.dj_helper_fn import fetch_nwb
+from ..utils.dj_helper_fn import fetch_nwb
 from ..common.common_nwbfile import AnalysisNwbfile
 from ..common.common_behav import RawPosition, VideoFile
 from ..common.common_interval import IntervalList
